dev-requirements.txt
numpy
pandas
scikit-learn
beautifulsoup4
selenium
wget
requests
langchain
langchain-core
langchain-community
langchain-openai
langchain-aws
langchain-chroma
langchain-qdrant
langgraph
qdrant-client
faiss-cpu
rank_bm25
openai
pypdf
pdfplumber
IPython
flashrank
streamlit